layout/base/tests/test_bug469170.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=469170
-->
<head>
  <title>Test for Bug 469170</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="runTest();">
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=469170">Mozilla Bug 469170</a></p>

<iframe id="source" width="50" height="50" srcdoc="<html></html>"></iframe>

<pre id="test">
<script type="application/javascript">

/** Test for Bug 469170 **/

SimpleTest.waitForExplicitFinish();

function runTest() {
  var source = document.getElementById('source').contentWindow;
  rect = { left: 0, top: 0,
           width: source.innerWidth, height: source.innerHeight };
  var canvas = SpecialPowers.snapshotRect(source, rect, "transparent");
  var context = canvas.getContext("2d");

  var components = [ "red", "green", "blue", "alpha" ];

  var data = context.getImageData(0, 0, canvas.width, canvas.height).data;
  var failed = false;
  for (var i = 0; i < data.length; i++) {
    if (data[i] != 0) {
      is(data[i], 0, "pixel " + Math.floor(i/4) + " " + components[i%4]);
      failed = true;
    }
  }
  if (!failed) {
    ok(!failed, "all pixels fully transparent");
  }
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
